# backend/openapi.yaml
openapi: 3.0.0
info:
  title: ProductAPI
  description: >-
    This is the API documentation for the application handling products and
    ransomware data.
  version: 1.0.0
  contact:
    name: Dev Team
    email: makwana.jignesh@gmail.com
...
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      operationId: auth-refresh
      tags:
        - auth
      summary: Token refresh
      description: Authenticates a user and returns JWT access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                refreshToken:
                  type: string
                  description: access token
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    type: string
                    description: access token
          headers:
            x-rate-limit:
              description: The maximum number of allowed requests per hour
              schema:
                type: integer
            x-rate-limit-remaining:
              description: The number of remaining requests in the current time window
              schema:
                type: integer
            x-rate-limit-reset:
              description: The time at which the rate limit will reset
              schema:
                type: integer
        '401':
          description: Invalid credentials
          headers:
            x-rate-limit:
              description: The maximum number of allowed requests per hour
              schema:
                type: integer
            x-rate-limit-remaining:
              description: The number of remaining requests in the current time window
              schema:
                type: integer
            x-rate-limit-reset:
              description: The time at which the rate limit will reset
              schema:
                type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...